+++ packages/keyring-api/src/api/address.ts
@@ -1,5 +1,5 @@
+import { object } from '@metamask/keyring-utils';
 import type { Infer } from '@metamask/superstruct';
-import { object } from '@metamask/superstruct';
 import { CaipAccountIdStruct } from '@metamask/utils';
 
 /**