+++ packages/keyring-api/src/v2/api/create-account/bip44.ts
@@ -1,4 +1,5 @@
-import { literal, number, object, string } from '@metamask/superstruct';
+import { object } from '@metamask/keyring-utils';
+import { literal, number, string } from '@metamask/superstruct';
 import type { Infer } from '@metamask/superstruct';
 
 import { DerivationPathStruct } from '../../../api/derivation';