+++ packages/keyring-api/src/v2/api/export-account/private-key.ts
@@ -1,4 +1,5 @@
-import { literal, object, string } from '@metamask/superstruct';
+import { object, sensitive } from '@metamask/keyring-utils';
+import { literal, string } from '@metamask/superstruct';
 import type { Infer } from '@metamask/superstruct';
 
 import { PrivateKeyEncodingStruct } from '../private-key';
@@ -14,7 +15,7 @@ export const PrivateKeyExportedAccountStruct = object({
   /**
    * The private key of the exported account.
    */
-  privateKey: string(),
+  privateKey: sensitive(string()),
   /**
    * The encoding of the exported private key.
    */