+++ packages/keyring-api/src/v2/api/keyring-capabilities.ts
@@ -1,11 +1,5 @@
-import {
-  array,
-  boolean,
-  exactOptional,
-  nonempty,
-  object,
-  partial,
-} from '@metamask/superstruct';
+import { exactOptional, object } from '@metamask/keyring-utils';
+import { array, boolean, nonempty, partial } from '@metamask/superstruct';
 import type { Infer } from '@metamask/superstruct';
 
 import { CaipChainIdStruct } from '../../api/caip';