+++ packages/keyring-api/src/v2/api/private-key.ts
@@ -1,4 +1,5 @@
-import { enums, exactOptional, object } from '@metamask/superstruct';
+import { exactOptional, object } from '@metamask/keyring-utils';
+import { enums } from '@metamask/superstruct';
 import type { Infer } from '@metamask/superstruct';
 
 import { KeyringAccountTypeStruct } from '../../api/account';